Add Original Artist of tracks to StationMedia entity
- Adds original_artist to the StationMedia entity.
- Adds original_artist to the Song API object and display with API.
- original_artist is a column in the database's station_media table.

This is the groundwork for a future enhancement I wish to make to the
Queue.php duplicate prevention code to allow different compositions
of tracks with the same title, where they are not covers, to be played.
I can set and display the original_artist with the NowPlaying endpoint,
but not the StationQueue, SongHistory and StationMedia (file) endpoints.